Update root CLAUDE.md
So I started with these rules from an article. I think a lot of these
things should be repo specific, and claude has been really overtesting
in a way I don't like.

Overtesting is almost as antisocial as no tests, let's calm things down
yeah. We have to remember the suffering of those that have to read our
PRs

It's getting there. I think claude as a tool is developing, so less is
more too in terms of your root config.